Marketing for Libraries Requests for New Info.
Hello,
This optional description is a quest for marketing strategies and plans, as part of a research study as a graduate student with a Bachelors of Science in Psychology, seeking Marketing Plans, as per the Crash Course book does not allow for a person to use anything but the contents of the book for files such as a commit file, or a master branch, including a code file. Please comment using github with any Marketing Plans you wish to deliver to Gabrielle M. Udero, you do not need the full submission, an example would be great! 

Thank you, 

Gabrielle Udero 
Sole Proprietor